Add files via upload
Choose the most suitable evaluation metric and state why you chose it.
Build a benchmark model for the given dataset.
Build a K-Nearest neighbours’ model for the given dataset and find the best value of K.
Build a Linear model for the given dataset with regularisation. Attempt to interpret the variable coefficients of the Linear Model.
Build a Decision tree model for the given dataset. Attempt to interpret the variable importance.
Plot the following Bar plots:
Train score of all the above models.
Test (not validation!) score of all the above models
Attempt to explain the observations from the plots (optional)